fix: refactor test directories and create ci job
