+++ tsconfig.json
@@ -19,6 +19,6 @@
   },
   "include": [
     "**/*.ts",
-    "e2e"
+    "tests"
   ]
 }
